# Optimize HPACK usage to align more with Netty types and remove heavy …