--- a/codec-http2/src/main/java/io/netty/handler/codec/http2/CharSequenceMap.java
+++ b/codec-http2/src/main/java/io/netty/handler/codec/http2/CharSequenceMap.java
@@ -18,11 +18,16 @@ package io.netty.handler.codec.http2;
 import io.netty.handler.codec.DefaultHeaders;
 import io.netty.handler.codec.UnsupportedValueConverter;
 import io.netty.handler.codec.ValueConverter;
+import io.netty.util.internal.UnstableApi;
 
 import static io.netty.util.AsciiString.CASE_INSENSITIVE_HASHER;
 import static io.netty.util.AsciiString.CASE_SENSITIVE_HASHER;
 
-final class CharSequenceMap<V> extends DefaultHeaders<CharSequence, V, CharSequenceMap<V>> {
+/**
+ * Internal use only!
+ */
+@UnstableApi
+public final class CharSequenceMap<V> extends DefaultHeaders<CharSequence, V, CharSequenceMap<V>> {
     public CharSequenceMap() {
         this(true);
     }
@@ -33,5 +38,11 @@ final class CharSequenceMap<V> extends DefaultHeaders<CharSequence, V, CharSeque
 
     public CharSequenceMap(boolean caseSensitive, ValueConverter<V> valueConverter) {
         super(caseSensitive ? CASE_SENSITIVE_HASHER : CASE_INSENSITIVE_HASHER, valueConverter);
+    }
+
+    @SuppressWarnings("unchecked")
+    public CharSequenceMap(boolean caseSensitive, ValueConverter<V> valueConverter, int arraySizeHint) {
+        super(caseSensitive ? CASE_SENSITIVE_HASHER : CASE_INSENSITIVE_HASHER, valueConverter,
+                NameValidator.NOT_NULL, arraySizeHint);
     }
 }