--- a/codec-http2/src/main/java/io/netty/handler/codec/http2/DefaultHttp2HeadersEncoder.java
+++ b/codec-http2/src/main/java/io/netty/handler/codec/http2/DefaultHttp2HeadersEncoder.java
@@ -16,19 +16,14 @@
 package io.netty.handler.codec.http2;
 
 import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufOutputStream;
+import io.netty.buffer.Unpooled;
 import io.netty.handler.codec.http2.internal.hpack.Encoder;
-import io.netty.util.AsciiString;
 import io.netty.util.internal.UnstableApi;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.util.Map.Entry;
 
 import static io.netty.handler.codec.http2.Http2CodecUtil.DEFAULT_HEADER_TABLE_SIZE;
 import static io.netty.handler.codec.http2.Http2Error.COMPRESSION_ERROR;
-import static io.netty.handler.codec.http2.Http2Error.INTERNAL_ERROR;
 import static io.netty.handler.codec.http2.Http2Error.PROTOCOL_ERROR;
 import static io.netty.handler.codec.http2.Http2Exception.connectionError;
 import static io.netty.util.internal.ObjectUtil.checkNotNull;
@@ -36,9 +31,9 @@ import static io.netty.util.internal.ObjectUtil.checkNotNull;
 @UnstableApi
 public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2HeadersEncoder.Configuration {
     private final Encoder encoder;
-    private final ByteArrayOutputStream tableSizeChangeOutput = new ByteArrayOutputStream();
     private final SensitivityDetector sensitivityDetector;
     private final Http2HeaderTable headerTable;
+    private final ByteBuf tableSizeChangeOutput = Unpooled.buffer();
 
     public DefaultHttp2HeadersEncoder() {
         this(DEFAULT_HEADER_TABLE_SIZE, NEVER_SENSITIVE);
@@ -52,7 +47,6 @@ public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2Hea
 
     @Override
     public void encodeHeaders(Http2Headers headers, ByteBuf buffer) throws Http2Exception {
-        final OutputStream stream = new ByteBufOutputStream(buffer);
         try {
             if (headers.size() > headerTable.maxHeaderListSize()) {
                 throw connectionError(PROTOCOL_ERROR, "Number of headers (%d) exceeds maxHeaderListSize (%d)",
@@ -61,24 +55,18 @@ public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2Hea
 
             // If there was a change in the table size, serialize the output from the encoder
             // resulting from that change.
-            if (tableSizeChangeOutput.size() > 0) {
-                buffer.writeBytes(tableSizeChangeOutput.toByteArray());
-                tableSizeChangeOutput.reset();
+            if (tableSizeChangeOutput.isReadable()) {
+                buffer.writeBytes(tableSizeChangeOutput);
+                tableSizeChangeOutput.clear();
             }
 
             for (Entry<CharSequence, CharSequence> header : headers) {
-                encodeHeader(header.getKey(), header.getValue(), stream);
+                encodeHeader(buffer, header.getKey(), header.getValue());
             }
         } catch (Http2Exception e) {
             throw e;
         } catch (Throwable t) {
             throw connectionError(COMPRESSION_ERROR, t, "Failed encoding headers block: %s", t.getMessage());
-        } finally {
-            try {
-                stream.close();
-            } catch (IOException e) {
-                throw connectionError(INTERNAL_ERROR, e, e.getMessage());
-            }
         }
     }
 
@@ -92,13 +80,8 @@ public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2Hea
         return this;
     }
 
-    private byte[] toBytes(CharSequence chars) {
-        AsciiString aString = AsciiString.of(chars);
-        return aString.isEntireArrayUsed() ? aString.array() : aString.toByteArray();
-    }
-
-    private void encodeHeader(CharSequence key, CharSequence value, OutputStream stream) throws IOException {
-        encoder.encodeHeader(stream, toBytes(key), toBytes(value), sensitivityDetector.isSensitive(key, value));
+    private void encodeHeader(ByteBuf out, CharSequence key, CharSequence value) {
+        encoder.encodeHeader(out, key, value, sensitivityDetector.isSensitive(key, value));
     }
 
     /**
@@ -113,8 +96,6 @@ public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2Hea
             try {
                 // No headers should be emitted. If they are, we throw.
                 encoder.setMaxHeaderTableSize(tableSizeChangeOutput, max);
-            } catch (IOException e) {
-                throw new Http2Exception(COMPRESSION_ERROR, e.getMessage(), e);
             } catch (Throwable t) {
                 throw new Http2Exception(PROTOCOL_ERROR, t.getMessage(), t);
             }