--- a/codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/HpackUtil.java
+++ b/codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/HpackUtil.java
@@ -31,19 +31,17 @@
  */
 package io.netty.handler.codec.http2.internal.hpack;
 
-import java.nio.charset.Charset;
-
 final class HpackUtil {
     /**
      * A string compare that doesn't leak timing information.
      */
-    static boolean equals(byte[] s1, byte[] s2) {
-        if (s1.length != s2.length) {
+    static boolean equals(CharSequence s1, CharSequence s2) {
+        if (s1.length() != s2.length()) {
             return false;
         }
         char c = 0;
-        for (int i = 0; i < s1.length; i++) {
-            c |= s1[i] ^ s2[i];
+        for (int i = 0; i < s1.length(); i++) {
+            c |= s1.charAt(i) ^ s2.charAt(i);
         }
         return c == 0;
     }