… codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/Encoder.java
… codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/HeaderListener.java
… codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/Huffman.java
… codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/HuffmanDecoder.java
… codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/HuffmanEncoder.java
… codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/StaticTable.java
… codec-http2/src/test/java/io/netty/handler/codec/http2/DefaultHttp2HeadersDecoderTest.java
… codec-http2/src/test/java/io/netty/handler/codec/http2/internal/hpack/DecoderTest.java
… codec-http2/src/test/java/io/netty/handler/codec/http2/internal/hpack/HuffmanTest.java
… codec-http2/src/test/java/io/netty/handler/codec/http2/internal/hpack/TestCase.java
… codec-http2/src/test/java/io/netty/handler/codec/http2/internal/hpack/TestHeaderListener.java
… microbench/src/main/java/io/netty/microbench/http2/internal/hpack/DecoderBenchmark.java
… microbench/src/main/java/io/netty/microbench/http2/internal/hpack/Header.java
… microbench/src/main/java/io/netty/microbench/http2/internal/hpack/HeadersSize.java