@nmittler 2016-06-06
@normanmaurer took a first pass ... looks great!

@carl-mastrangelo would you mind trying this out to see how well this addresses the garbage problem?