@Scottmitch 2016-06-06
@normanmaurer - Plz add https://github.com/netty/netty/issues/3597 to the commit message too. This commit partially resolves the issues described in that issue.