@normanmaurer 2016-06-06
@Scottmitch did I miss anything ? I have https://github.com/netty/netty/issues/3597 in the commit message.