@Scottmitch 2016-06-06
@normanmaurer - just https://github.com/netty/netty/pull/5355#discussion_r65934422 left to discuss  ... the same comment would apply to `HuffmanEncoder` too. Everything else lgtm