@normanmaurer 2016-06-08
@carl-mastrangelo @nmittler @Scottmitch @buchgr PTAL again... So the thing is this HPACK implementation is only "compliant" in the scope of HTTP2 as it expect its name / value to be ascii. This is good enough for our use-case and allows us to do some optimisations.