@nmittler 2016-06-08
@normanmaurer changes LGTM - great job! Can you post some benchmarks to show before and after this change to make sure there is no regression?