@normanmaurer 2016-06-08
@nmittler yeah will need to write some... Can you also run the grpc benchmarks with this ?