@nmittler 2016-06-08
@normanmaurer yep, good idea.

@carl-mastrangelo could you benchmark this with grpc?