@carl-mastrangelo 2016-06-08
Allocation Count is way better now.  Looking At allocation count for turning gRPC Headers into actual buffered bytes, here are all the allocations:

Pseudo code:
```
      enabled = true;
      m = new Metadata();
      Http2Headers heads = io.grpc.netty.Utils.convertServerHeaders(m);
      enc.encodeHeaders(heads, buf1);
      enabled = false;
```

Before:
```
I just allocated the object {} of type java/util/LinkedHashMap whose size is 56
I just allocated the object Metadata({}) of type io/grpc/Metadata whose size is 24
I just allocated the object [[B@142269f2 of type [B whose size is 16
  It's an array of size 0
I just allocated the object java.util.LinkedHashMap$LinkedEntryIterator@331acdad of type java/util/LinkedHashMap$LinkedEntryIterator whose size is 32
I just allocated the object [Lio.netty.util.AsciiString;@41d426b5 of type io/netty/util/AsciiString whose size is 32
  It's an array of size 4
I just allocated the object [Lio.netty.util.AsciiString;@8dbfffb of type io/netty/util/AsciiString whose size is 16
  It's an array of size 0
I just allocated the object io.grpc.netty.GrpcHttp2Headers@f316aeb of type io/grpc/netty/GrpcHttp2Headers whose size is 24
I just allocated the object [B@6aa3a905 of type byte whose size is 24
  It's an array of size 8
I just allocated the object java.io.DataOutputStream@a22cb6a of type java/io/DataOutputStream whose size is 32
I just allocated the object io.netty.buffer.ByteBufOutputStream@5dd1c9f2 of type io/netty/buffer/ByteBufOutputStream whose size is 24
I just allocated the object io.grpc.netty.GrpcHttp2Headers$Itr@59b38691 of type io/grpc/netty/GrpcHttp2Headers$Itr whose size is 32
I just allocated the object sun.nio.cs.ISO_8859_1$Decoder@2e2ff723 of type sun/nio/cs/ISO_8859_1$Decoder whose size is 40
I just allocated the object [C@324a0017 of type char whose size is 32
  It's an array of size 7
I just allocated the object :status of type java/lang/String whose size is 24
88be

```

After:
```
I just allocated the object {} of type java/util/LinkedHashMap whose size is 56
I just allocated the object Metadata({}) of type io/grpc/Metadata whose size is 24
I just allocated the object [[B@8dbfffb of type [B whose size is 16
  It's an array of size 0
I just allocated the object java.util.LinkedHashMap$LinkedEntryIterator@f316aeb of type java/util/LinkedHashMap$LinkedEntryIterator whose size is 32
I just allocated the object [Lio.netty.util.AsciiString;@6aa3a905 of type io/netty/util/AsciiString whose size is 32
  It's an array of size 4
I just allocated the object [Lio.netty.util.AsciiString;@a22cb6a of type io/netty/util/AsciiString whose size is 16
  It's an array of size 0
I just allocated the object io.grpc.netty.GrpcHttp2Headers@5dd1c9f2 of type io/grpc/netty/GrpcHttp2Headers whose size is 24
I just allocated the object io.grpc.netty.GrpcHttp2Headers$Itr@59b38691 of type io/grpc/netty/GrpcHttp2Headers$Itr whose size is 32
88be
```