@carl-mastrangelo 2016-06-09
Running the benchmarks causes less enthusiasm.  The tests are almost 100% noise.  I don't have numbers to tell if this is actually faster or not.