@carl-mastrangelo 2016-06-09
Re running shows clearer but worse results:

Before:
```
Benchmark                              (headerCount)    Mode     Cnt     Score    Error  Units
HeadersBenchmark.convertClientHeaders              1  sample  128784   632.251 ±  3.357  ns/op
HeadersBenchmark.convertClientHeaders              5  sample  166694   965.962 ± 47.415  ns/op
HeadersBenchmark.convertClientHeaders             10  sample  111651  1403.191 ±  5.374  ns/op
HeadersBenchmark.convertClientHeaders             20  sample  136152  2305.764 ± 54.399  ns/op
HeadersBenchmark.convertServerHeaders              1  sample  120919   370.659 ± 67.219  ns/op
HeadersBenchmark.convertServerHeaders              5  sample  113165   707.157 ±  3.854  ns/op
HeadersBenchmark.convertServerHeaders             10  sample  136208  1176.250 ± 60.380  ns/op
HeadersBenchmark.convertServerHeaders             20  sample  147716  2171.886 ± 93.689  ns/op
```

After:
```
Benchmark                              (headerCount)    Mode     Cnt     Score    Error  Units
HeadersBenchmark.convertClientHeaders              1  sample  129991   641.987 ±  2.989  ns/op
HeadersBenchmark.convertClientHeaders              5  sample  151857  1081.565 ± 52.067  ns/op
HeadersBenchmark.convertClientHeaders             10  sample  188208  1689.725 ± 40.720  ns/op
HeadersBenchmark.convertClientHeaders             20  sample  109098  2883.036 ± 69.334  ns/op
HeadersBenchmark.convertServerHeaders              1  sample  131745   314.715 ±  2.454  ns/op
HeadersBenchmark.convertServerHeaders              5  sample  107129   752.784 ±  4.652  ns/op
HeadersBenchmark.convertServerHeaders             10  sample  124304  1307.933 ± 87.235  ns/op
HeadersBenchmark.convertServerHeaders             20  sample  133735  2389.373 ± 73.400  ns/op
```